move _resource.pyx to outer dir
