+++ mars/services/scheduling/supervisor/queueing.py
@@ -22,8 +22,8 @@ from typing import DefaultDict, Dict, List, Optional, Tuple, Union
 
 from .... import oscar as mo
 from ....lib.aio import alru_cache
+from ....resource import Resource
 from ....utils import dataslots
-from ... import Resource
 from ...subtask import Subtask
 from ...task import TaskAPI
 from ..utils import redirect_subtask_errors
